tools/requirements.txt
PyMuPDF>=1.19.0
python-docx>=0.8.11
argparse>=1.4.0
pytesseract>=0.3.8
Pillow>=8.0.0
pikepdf>=2.0.0
tqdm>=4.62.3
pandas>=1.3.0
openpyxl>=3.0.7
pdf2image
ocrmypdf
langdetect
easyocr
torch 
torchvision
pydub
numpy
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib

fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
urllib3==1.26.15
bcrypt==4.3.0
python-telegram-bot==22.0
gdown==5.2.0
docx2pdf==0.1.8